+++ __tests__/index.spec.js
@@ -5,6 +5,8 @@ function validateRegularBob(bob) {
   expect(bob.country).toBeUndefined()
   expect(bob.aliases).toBeUndefined()
   expect(bob.arrayWithDefault).toBeUndefined()
+  expect(bob.nestedWithoutDefaults).toBeUndefined()
+  expect(bob.defaultWithSchemaTypeFn).toBeUndefined()
   expect(bob.nested).toBeUndefined()
 }
 function validateBob(bob) {
@@ -12,21 +14,27 @@ function validateBob(bob) {
   expect(bob.aliases).toBeDefined()
   expect(bob.aliases).toHaveLength(0)
   expect(bob.arrayWithDefault).toBeUndefined()
+  expect(bob.nestedWithoutDefaults).toBeDefined()
+  expect(bob.nestedWithoutDefaults.test).toBeUndefined()
   expect(bob.nested).toBeDefined()
   expect(bob.nested.prop).toBe('Default')
   expect(bob.nested.other).toBe(true)
   expect(bob.nested.noDefault).toBeUndefined()
   expect(bob.fnDefault).toBe('Bob')
+  expect(bob.defaultWithSchemaTypeFn).toBe('Test Default')
 }
 function validateAlice(alice) {
   expect(alice.country).toBe('CA')
   expect(alice.aliases).toBeDefined()
   expect(alice.arrayWithDefault).toBeUndefined()
+  expect(alice.nestedWithoutDefaults).toBeDefined()
+  expect(alice.nestedWithoutDefaults.test).toBe('test')
   expect(alice.nested).toBeDefined()
   expect(alice.nested.prop).toBe('Prop')
   expect(alice.nested.other).toBe(false)
   expect(alice.nested.noDefault).toBe('Test')
   expect(alice.fnDefault).toBe('Alice')
+  expect(alice.defaultWithSchemaTypeFn).toBe('Value')
 }
 
 const bobId = '5d23fa87d7f8b00011fa25c5'
@@ -44,6 +52,8 @@ describe('mongooseLeanDefaults', () => {
       name: String
     }, { collection: 'users' })
     const OldUser = mongoose.model('OldUser', oldSchema)
+    // ensure bob is not on database
+    await OldUser.deleteOne({ _id: bobId })
     await OldUser.create({
       _id: bobId,
       name: 'Bob'
@@ -53,6 +63,10 @@ describe('mongooseLeanDefaults', () => {
       name: String,
       country: { type: String, default: 'USA' },
       aliases: [String],
+      nestedWithoutDefaults: {
+        test: String,
+      },
+      defaultWithSchemaTypeFn: String,
       arrayWithDefault: { type: [String], default: undefined },
       nested: {
         prop: {
@@ -65,20 +79,27 @@ describe('mongooseLeanDefaults', () => {
         },
         noDefault: String
       },
-      fnDefault: { type: String, default: function() { return this.name; } }
+      fnDefault: { type: String, default: function () { return this.name } }
     }, { collection: 'users' })
+    schema.path('defaultWithSchemaTypeFn').default('Test Default')
     schema.plugin(mongooseLeanDefaults)
     User = mongoose.model('User', schema)
+    // ensure alice is not on database
+    await User.deleteOne({_id: aliceId})
     await User.create({
       _id: aliceId,
       name: 'Alice',
       country: 'CA',
       aliases: ['Ally'],
+      nestedWithoutDefaults: {
+        test: 'test'
+      },
       nested: {
         prop: 'Prop',
         other: false,
         noDefault: 'Test'
-      }
+      },
+      defaultWithSchemaTypeFn: 'Value'
     })
     done()
   })